% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_episodeExplorerDataInput.R
\name{mod_episodeExplorerData}
\alias{mod_episodeExplorerData}
\title{mod_episodeExplorerData server function}
\usage{
mod_episodeExplorerData(
  input,
  output,
  session,
  lang,
  year_explorer_output_reactives,
  parent_session,
  cache
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}

\item{year_explorer_output_reactives}{reactives from year report map}

\item{parent_session}{parent session to update the active tab}

\item{cache}{cache for storing the episode sel value}
}
\description{
mod_episodeExplorerData server function
}
