% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_episodeExplorerDataInput.R
\name{mod_episodeExplorerDataInput}
\alias{mod_episodeExplorerDataInput}
\title{mod_episodeExplorerDataInput and mod_episodeExplorerData}
\usage{
mod_episodeExplorerDataInput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the data inputs
}
