% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_episodeExplorerPlotOutput.R
\name{mod_episodeExplorerPlot}
\alias{mod_episodeExplorerPlot}
\title{mod_episodeExplorerplot server function}
\usage{
mod_episodeExplorerPlot(
  input,
  output,
  session,
  lang,
  episode_explorer_data_reactives
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}

\item{episode_explorer_data_reactives}{data reactive}
}
\description{
mod_episodeExplorerplot server function
}
