% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_episodeExplorerPlotOutput.R
\name{mod_episodeExplorerPlotOutput}
\alias{mod_episodeExplorerPlotOutput}
\title{mod_episodeExplorerPlotOutput and mod_episodeExplorerPlot}
\usage{
mod_episodeExplorerPlotOutput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the data outputs
}
