% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infoOutput.R
\name{mod_info}
\alias{mod_info}
\title{mod_info server function}
\usage{
mod_info(
  input,
  output,
  session,
  lang,
  year_explorer_data_reactives,
  year_explorer_output_reactives,
  trigger
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}

\item{year_explorer_data_reactives, year_explorer_output_reactives}{data reactives}

\item{trigger}{character with the trigger (species_id or county_name)}
}
\description{
mod_info server function
}
