% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infoOutput.R
\name{mod_infoOutput}
\alias{mod_infoOutput}
\title{mod_infoOutput and mod_info}
\usage{
mod_infoOutput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
module for creating the astounding viz when click
}
