% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_saveDataUI.R
\name{mod_saveDataUI}
\alias{mod_saveDataUI}
\alias{mod_saveData}
\title{mod_saveDataUI and mod_saveData}
\usage{
mod_saveDataUI(id)

mod_saveData(input, output, session, lang, data_reactives, data_type)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{language selected}

\item{data_reactives}{reactives with the needed data to save}

\item{data_type}{character vector of length 1, with the type of data to save: "episode_explorer" or
"year_explorer"}
}
\description{
module for creating the astounding viz when click
}
