% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_speciesInfoOutput.R
\name{mod_speciesInfo}
\alias{mod_speciesInfo}
\title{mod_info server function}
\usage{
mod_speciesInfo(
  input,
  output,
  session,
  lang,
  year_explorer_data_reactives,
  year_explorer_output_reactives
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}

\item{year_explorer_data_reactives, year_explorer_output_reactives}{data reactives}
}
\description{
mod_info server function
}
