% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_speciesInfoOutput.R
\name{mod_speciesInfoOutput}
\alias{mod_speciesInfoOutput}
\title{mod_speciesInfoOutput and mod_speciesInfo}
\usage{
mod_speciesInfoOutput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
module for creating the astounding viz when click
}
