% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_yearExplorerOutput.R
\name{mod_yearExplorer}
\alias{mod_yearExplorer}
\title{mod_yearExplorer server function}
\usage{
mod_yearExplorer(input, output, session, lang, year_explorer_data_reactives)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}

\item{year_explorer_data_reactives}{data reactive}
}
\description{
mod_yearExplorer server function
}
