% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_yearExplorerDataInput.R
\name{mod_yearExplorerData}
\alias{mod_yearExplorerData}
\title{mod_yearExplorerData server function}
\usage{
mod_yearExplorerData(input, output, session, lang, cache)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive}

\item{cache}{memoryCache object to store the selected inputs and retrieve them when the underlying data
changes.}
}
\description{
mod_yearExplorerData server function
}
