% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_yearExplorerDataInput.R
\name{mod_yearExplorerDataInput}
\alias{mod_yearExplorerDataInput}
\title{mod_yearExplorerDataInput and mod_yearExplorerData}
\usage{
mod_yearExplorerDataInput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the data inputs
}
