% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_yearExplorerInfoOutput.R
\name{mod_yearExplorerInfoOutput}
\alias{mod_yearExplorerInfoOutput}
\title{mod_yearExplorerInfoOutput and mod_yearExplorerInfo}
\usage{
mod_yearExplorerInfoOutput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
module for creating the astounding viz when click
}
