% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_yearExplorerOutput.R
\name{mod_yearExplorerOutput}
\alias{mod_yearExplorerOutput}
\title{mod_yearExplorerOutput and mod_yearExplorer}
\usage{
mod_yearExplorerOutput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the data outputs
}
